lambda/athlete_upload/requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
requests==2.31.0
soupsieve==2.5
typing_extensions==4.9.0
urllib3==2.1.0
aws-psycopg2
